Update thrift files (#166)
* Update thirift files

* Revert port in the example

* fix testcase

+++ result_set.go
@@ -335,7 +335,7 @@ func (res ResultSet) GetErrorCode() ErrorCode {
 	return ErrorCode(res.resp.ErrorCode)
 }
 
-func (res ResultSet) GetLatency() int32 {
+func (res ResultSet) GetLatency() int64 {
 	return res.resp.LatencyInUs
 }
