+++ result_set_test.go
@@ -550,7 +550,7 @@ func TestResultSet(t *testing.T) {
 	}
 
 	assert.Equal(t, ErrorCode_E_STATEMENT_EMPTY, resultSetWithNil.GetErrorCode())
-	assert.Equal(t, int32(1000), resultSetWithNil.GetLatency())
+	assert.Equal(t, int64(1000), resultSetWithNil.GetLatency())
 	assert.Equal(t, "", resultSetWithNil.GetErrorMsg())
 	assert.Equal(t, "", resultSetWithNil.GetSpaceName())
 	assert.Equal(t, "", resultSetWithNil.GetComment())
@@ -594,7 +594,7 @@ func TestResultSet(t *testing.T) {
 		t.Error(err)
 	}
 	assert.Equal(t, ErrorCode_SUCCEEDED, resultSet.GetErrorCode())
-	assert.Equal(t, int32(1000), resultSet.GetLatency())
+	assert.Equal(t, int64(1000), resultSet.GetLatency())
 	assert.Equal(t, "test_err_msg", resultSet.GetErrorMsg())
 	assert.Equal(t, "test_space", resultSet.GetSpaceName())
 	assert.Equal(t, "test_comment", resultSet.GetComment())
